sound: soc: qdsp6v2: add missing size multiplier to memsets
Currenty updt_params is of type uint32_t which is 4 byte in size.
MAX_ENABLE_CMD_SIZE is currently defined as 32.

updt_params has an array size of 32 and is of type uint32_t[].
That means the total size of it is 32*4 which equals to 128.

The previous memset statements were zeroing out the first 32 byte
of updt_params instead of the 128 which is result of the max enabled
cmd size (32 byte) multiplied by the size of uint32_t (4 byte).

Not correcting this would theoretically be able to lead to data leaks,
maybe in a future point of time because if anything is written beyond
the 32 byte (MAX_ENABLED_CMD_SIZE) and then the first 32 byte are
zeroed out, everything after the first 32 byte can still be read.

This also fixes the warning -Wmemset-elt-size

Change-Id: If16612be757245137796175f543b0178e9026721
Signed-off-by: Simao Gomes Viana <xdevs23@outlook.com>